Cuadra_Hamzah/PS4/requirements.txt
flask==3.0.3
mysql-connector==2.2.9